+++ relm4-macros/tests/ui/compile-fail/no-view-output.rs
@@ -15,12 +15,12 @@ impl SimpleComponent for AppModel {
     #[allow(unreachable_code)]
     fn init(
         _: (),
-        root: Self::Root,
-        sender: ComponentSender<Self>,
+        _root: Self::Root,
+        _sender: ComponentSender<Self>,
     ) -> ComponentParts<Self> {
         let model = Self;
 
-        let widgets = todo!();
+        let widgets = AppModelWidgets {};
 
         ComponentParts { model, widgets }
     }