+++ relm4/src/component/sync/builder.rs
@@ -14,7 +14,6 @@ use std::any;
 use std::cell::RefCell;
 use std::marker::PhantomData;
 use std::rc::Rc;
-use tokio::sync::oneshot;
 use tracing::info_span;
 
 /// A component that is ready for docking and launch.
@@ -183,8 +182,6 @@ impl<C: Component> ComponentBuilder<C> {
         // Gets notifications when a component's model and view is updated externally.
         let (notifier, notifier_receiver) = crate::channel();
 
-        let (source_id_sender, source_id_receiver) = oneshot::channel::<glib::JoinHandle<()>>();
-
         // Encapsulates the senders used by component methods.
         let component_sender = ComponentSender::new(
             input_sender.clone(),
@@ -211,8 +208,7 @@ impl<C: Component> ComponentBuilder<C> {
         // Spawns the component's service. It will receive both `Self::Input` and
         // `Self::CommandOutput` messages. It will spawn commands as requested by
         // updates, and send `Self::Output` messages externally.
-        let handle = crate::spawn_local_with_priority(priority, async move {
-            let id = source_id_receiver.await.unwrap().into_source_id().unwrap();
+        crate::spawn_local_with_priority(priority, async move {
             let mut notifier = GuardedReceiver::new(notifier_receiver);
             let mut cmd = GuardedReceiver::new(cmd_receiver);
             let mut input = GuardedReceiver::new(input_receiver);
@@ -276,15 +272,11 @@ impl<C: Component> ComponentBuilder<C> {
 
                         shutdown_notifier.shutdown();
 
-                        id.remove();
-
                         return;
                     }
                 );
             }
         });
-
-        source_id_sender.send(handle).unwrap();
 
         // Give back a type for controlling the component service.
         Connector {