+++ CHANGES.md
@@ -6,6 +6,10 @@
 
 + core: Add `allow_multiple_instances()` to `RelmApp` to allow multiple concurrent instances of the same application
 
+### Changed
+
++ core: Simplified internal code for runtime creation
+
 ### Fixed
 
 + core: Don't require `Clone` and `Debug` for the generic action name parameter in `RelmAction`