+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 ### Added
 
++ core: Add `set_align` and `set_expand` to `RelmWidgetExt` to set horizontal and vertical properties at once
 + core: Add `allow_multiple_instances()` to `RelmApp` to allow multiple concurrent instances of the same application
 + core: Add extension traits for `adw::PreferencesPage`